Fix: replace non-standard 'uint' with 'size_t' in to_jpg.cpp
GCC 15 (shipped with ESP-IDF 6.0) with Picolibc does not provide 'uint'
as an implicit typedef. This causes a compilation error on line 191.

The rest of the codebase uses a local typedef in jpge.h, but to_jpg.cpp
does not include that header. Use size_t to match the type of max_len
which buf_size is assigned to.

+++ conversions/to_jpg.cpp
@@ -188,7 +188,7 @@ protected:
     size_t max_len, index;
 
 public:
-    memory_stream(void *pBuf, uint buf_size) : out_buf(static_cast<uint8_t*>(pBuf)), max_len(buf_size), index(0) { }
+    memory_stream(void *pBuf, size_t buf_size) : out_buf(static_cast<uint8_t*>(pBuf)), max_len(buf_size), index(0) { }
 
     virtual ~memory_stream() { }
